+++ packages/fabric/editor-core/src/nodeviews/ui/media-single/index.tsx
@@ -1,18 +1,27 @@
 import * as React from 'react';
 import { PureComponent, ReactElement } from 'react';
 import { Node as PMNode } from 'prosemirror-model';
-
+import { EditorView } from 'prosemirror-view';
 import { Wrapper } from './styled';
 import { MediaNodeProps } from '../media';
-import { Props as MediaProps } from '../../../ui/Media/MediaComponent';
+import { stateKey, MediaPluginState } from '../../../plugins/media';
 
 export interface MediaSingleNodeProps {
   node: PMNode;
+  view: EditorView;
 }
 
 export default class MediaSingleNode extends PureComponent<
   MediaSingleNodeProps
 > {
+  componentDidUpdate() {
+    const mediaPluginState: MediaPluginState = stateKey.getState(
+      this.props.view.state,
+    ) as MediaPluginState;
+    const { layout } = this.props.node.attrs;
+    mediaPluginState.updateLayout(layout);
+  }
+
   render() {
     const child: ReactElement<MediaNodeProps> = React.Children.only(
       React.Children.toArray(this.props.children)[0],
@@ -28,8 +37,8 @@ export default class MediaSingleNode extends PureComponent<
               width: '100%',
               height: '100%',
             },
-            appearance: 'image',
-          } as MediaProps,
+            isMediaSingle: true,
+          } as MediaNodeProps,
         )}
       </Wrapper>
     );