+++ packages/fabric/conversation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/conversation
 
+## 3.0.11
+
 ## 3.0.10
 - [patch] Fix dependencies in CI [670e930](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/670e930)