+++ packages/fabric/editor-core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/editor-core
 
+## 60.17.2
+
+## 60.17.1
+- [patch] validate incoming ADF node inserted from macro browser [e9d0af2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e9d0af2)
+
+## 60.17.0
+- [minor] Added floating toolbar to media single [46fdd15](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/46fdd15)
+
+## 60.16.1
+- [patch] Add support for single image wrap left/right layout [59d9a74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/59d9a74)
+
 ## 60.16.0
 - [minor] Enter keypress should always submit in stride. [51020fe](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/51020fe)